🎨 Move transform pose to own module
